pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "documind"
version = "0.1.0"
description = "Convert DOCX documents into LLM-oriented Markdown chunk sets."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-docx>=1.1.2",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.11",
  "pytest>=8.0",
]

[project.scripts]
documind = "documind.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
